backend/requirements/prod.txt
# 生产环境依赖模板

# 继承基础依赖
-r base.txt

# WSGI 服务器
gunicorn==21.2.0
whitenoise==6.6.0

# 监控和日志
sentry-sdk==1.38.0
newrelic==9.2.0
structlog==23.2.0

# 安全
django-ratelimit==4.1.0

# 性能优化
django-cachalot==2.6.1
django-compressor==4.4

# 健康检查
django-health-check==3.17.0

# 开发工具
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-silk==5.0.4

# 管理工具
django-admin-interface==0.25.0
